+++ tests/unit/services/test_bootstrap_race.py
@@ -1,11 +1,4 @@
-"""First-admin bootstrap must yield exactly one admin even under
-concurrent first-sign-ins.
-
-Two `ensure_user_row` calls fired with `asyncio.gather` against an empty
-DB: both observe `count_admins == 0`, both attempt to grant admin. The
-post-grant rollback (lexicographic tie-break) must demote the loser
-before either request returns.
-"""
+"""Default-role bootstrap must remain race-safe under concurrent sign-ins."""
 
 import asyncio
 import sys
@@ -29,7 +22,8 @@ from session_manager import User  # noqa: E402
 
 
 @pytest_asyncio.fixture
-async def session_factory():
+async def session_factory(monkeypatch):
+    monkeypatch.setenv("OPENRAG_DEFAULT_ROLE", "user")
     engine = create_async_engine(
         "sqlite+aiosqlite:///:memory:",
         connect_args={"check_same_thread": False},
@@ -45,13 +39,8 @@ async def session_factory():
 
 
 @pytest.mark.asyncio
-async def test_bootstrap_race_yields_single_admin(session_factory):
-    """Two concurrent first-sign-ins → exactly one admin.
-
-    Each call uses its own session (mirrors what FastAPI does per
-    request). The post-grant rollback must observe the second admin
-    and demote whichever caller is not min(user_id).
-    """
+async def test_bootstrap_race_assigns_default_role(session_factory):
+    """Two concurrent first-sign-ins both receive the default role."""
 
     async def signin(user_id: str) -> None:
         async with session_factory() as session:
@@ -66,23 +55,20 @@ async def test_bootstrap_race_yields_single_admin(session_factory):
             )
             await session.commit()
 
-    # SQLite serializes writes, so true parallelism is limited; but the
-    # service does several reads/writes per call, so even on SQLite the
-    # interleaving exercises the rollback branch.
     await asyncio.gather(signin("alice"), signin("bob"))
 
     async with session_factory() as session:
-        admins = await RoleRepo(session).list_admin_user_ids()
+        repo = RoleRepo(session)
+        alice_roles = await repo.list_user_roles("alice")
+        bob_roles = await repo.list_user_roles("bob")
 
-    assert len(admins) == 1, f"expected 1 admin, got {admins}"
-    # Lexicographic min wins
-    assert admins[0] == "alice"
+    assert {r.name for r in alice_roles} == {"user"}
+    assert {r.name for r in bob_roles} == {"user"}
 
 
 @pytest.mark.asyncio
-async def test_bootstrap_loser_falls_through_to_default_role(session_factory):
-    """The demoted bootstrap loser must still end up with the default
-    role, not zero roles."""
+async def test_concurrent_signins_do_not_grant_admin(session_factory):
+    """Concurrent first sign-ins must not grant admin implicitly."""
 
     async def signin(user_id: str) -> None:
         async with session_factory() as session:
@@ -101,19 +87,16 @@ async def test_bootstrap_loser_falls_through_to_default_role(session_factory):
 
     async with session_factory() as session:
         repo = RoleRepo(session)
-        admins = await repo.list_admin_user_ids()
+        alice_roles = await repo.list_user_roles("alice")
         zach_roles = await repo.list_user_roles("zach")
 
-    assert admins == ["alice"]
-    role_names = {r.name for r in zach_roles}
-    # Default role is "user" (set by OPENRAG_DEFAULT_ROLE, default "user")
-    assert "user" in role_names, f"loser should have default role, got {role_names}"
-    assert "admin" not in role_names
+    assert {r.name for r in alice_roles} == {"user"}
+    assert {r.name for r in zach_roles} == {"user"}
 
 
 @pytest.mark.asyncio
 async def test_no_race_single_signin_unchanged(session_factory):
-    """Sanity check — when there's no race, the first user becomes admin."""
+    """Sanity check — one sign-in gets the default role."""
     async with session_factory() as session:
         await ensure_user_row(
             session,
@@ -121,8 +104,8 @@ async def test_no_race_single_signin_unchanged(session_factory):
         )
         await session.commit()
     async with session_factory() as session:
-        admins = await RoleRepo(session).list_admin_user_ids()
-    assert admins == ["solo"]
+        roles = await RoleRepo(session).list_user_roles("solo")
+    assert {r.name for r in roles} == {"user"}
 
 
 @pytest.mark.asyncio
@@ -144,7 +127,7 @@ async def test_concurrent_signins_same_user_no_integrity_error(session_factory, 
 
     monkeypatch.setattr(_engine_mod, "SessionLocal", session_factory, raising=False)
 
-    from dependencies import _ENSURE_LOCKS, _ENSURED_USER_IDS, _ensure_db_user
+    from auth.user_identity_cache import _ENSURE_LOCKS, _ENSURED_USER_IDS, _ensure_db_user
 
     _ENSURED_USER_IDS.clear()
     _ENSURE_LOCKS.clear()
@@ -229,7 +212,7 @@ async def test_cache_keys_are_per_provider_subject_pair(session_factory, monkeyp
 
     monkeypatch.setattr(_engine_mod, "SessionLocal", session_factory, raising=False)
 
-    from dependencies import _ENSURE_LOCKS, _ENSURED_USER_IDS, _ensure_db_user
+    from auth.user_identity_cache import _ENSURE_LOCKS, _ENSURED_USER_IDS, _ensure_db_user
 
     _ENSURED_USER_IDS.clear()
     _ENSURE_LOCKS.clear()
@@ -274,7 +257,7 @@ async def test_invalidate_pops_only_target_identity(session_factory, monkeypatch
 
     monkeypatch.setattr(_engine_mod, "SessionLocal", session_factory, raising=False)
 
-    from dependencies import (
+    from auth.user_identity_cache import (
         _ENSURE_LOCKS,
         _ENSURED_USER_IDS,
         _ensure_db_user,
